exercises 4.20 - 4.23

+++ part4/blog-list/controllers/blogs.js
@@ -1,6 +1,5 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-const User = require('../models/user')
 const { verifyToken } = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
@@ -24,8 +23,7 @@ blogsRouter.post('/', verifyToken, async (request, response) => {
     return response.status(400).json({ error: 'title missing' })
   }
 
-  const user = await User.findById(request.authUserId)
-
+  const user = request.authUser
   const blog = new Blog({ ...request.body, user: user.id })
   const savedBlog = await blog.save()
 
@@ -35,13 +33,22 @@ blogsRouter.post('/', verifyToken, async (request, response) => {
   response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', verifyToken, async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+  if (!blog) {
+    return response.status(204).json()
+  }
+
+  if (request.authUser._id.toString() !== blog.user.toString()) {
+    return response.status(403).json({ error: 'Forbidden' })
+  }
+
   const deletedBlog = await Blog.findByIdAndDelete(request.params.id)
 
   response.status(204).json(deletedBlog)
 })
 
-blogsRouter.put('/:id', async (request, response) => {
+blogsRouter.put('/:id', verifyToken, async (request, response) => {
   const body = request.body
   if (!body.url) {
     return response.status(400).json({ error: 'url missing' })
@@ -53,6 +60,15 @@ blogsRouter.put('/:id', async (request, response) => {
 
   if (!body.title) {
     return response.status(400).json({ error: 'title missing' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+  if (!blog) {
+    return response.status(404)
+  }
+
+  if (request.authUser._id.toString() !== blog.user.toString()) {
+    return response.status(403).json({ error: 'Forbidden' })
   }
 
   const updatedBlog = await Blog.findByIdAndUpdate(
